Added a small lock (test-1-passed)

+++ main.py
@@ -332,9 +332,13 @@ class Moderation(commands.Cog):
 
     @commands.command(name='block', help='Blocks a user from chatting in a specific channel.')
     async def block(self, ctx, member: discord.Member, *, reason='No reason provided.'):
-        await ctx.channel.set_permissions(member, send_messages=False)
-        await ctx.message.delete()
-        await ctx.send(f'You\'re now blocked from chatting, {member.mention} | Reason: {reason}')
+        if member != ctx.author:
+            await ctx.channel.set_permissions(member, send_messages=False)
+            await ctx.message.delete()
+            await ctx.send(f'You\'re now blocked from chatting, {member.mention} | Reason: {reason}')
+
+        else:
+            await ctx.send(f'You can\'t block yourself!')
 
     @commands.command(name='unblock', help='Unblocks a user.')
     async def unblock(self, ctx, member: discord.Member):
